added user authentication in blogs app

+++ part4/blog list/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogListRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogListRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -11,19 +12,23 @@ blogListRouter.get('/', async (request, response) => {
 })
 
 blogListRouter.post('/', async (request, response) => {
+    if(!request.token)
+    {
+        return response.status(400).send({error : 'no token provided'})
+    }   
     if (!request.body.title || !request.body.url) {
         return response.status(400).end()
     }
     if (!request.body.likes) {
         request.body.likes = 0
     }
+    const user= request.user
     const blog = new Blog(request.body)
-    const result = await blog.save()
-    const user = await User.findById(request.body.user) 
+    blog.user = user._id
+    const result = await blog.save()    
     user.blogs = user.blogs.concat(result.id)
     await user.save()
     response.status(201).json(result)
-
 })
 
 blogListRouter.get('/:id', async (request, response) => {
@@ -38,7 +43,24 @@ blogListRouter.get('/:id', async (request, response) => {
 })
 
 blogListRouter.delete('/:id', async (request, response) => {
+    if(!request.token)
+    {
+        return response.status(400).send({error : 'no token provided'})
+    }   
+
+    const blog = await Blog.findById(request.params.id)
+    const user= request.user
+    if(!blog){
+        return response.status(404).send({error : 'no blog found with given id'})
+    } 
+
+    if((blog.user.toString() !== user.id)){
+        return response.status(401).send({error : 'user unauthorized to delete'})
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(b => b.toString()!==request.params.id)
+    await user.save()
     response.status(204).end()
 })
